backend/requirements.txt
fastapi>=0.135.0
uvicorn>=0.44.0
python-multipart>=0.0.26
sqlalchemy>=2.0.0
aiosqlite>=0.22.0
chromadb>=1.5.0
langchain-text-splitters>=1.1.0
openai>=2.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pypdf>=6.0.0
python-docx>=1.0.0
